Material/14_day/14_2_day/pyproject.toml
[project]
name = "14-2-day"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.0",
    "mysql-connector-python>=9.3.0",
    "sql>=2022.4.0",
]
